Added callback functions book CRUD functionality.
Previously 'save-new', 'save' and 'delete' buttons event handlers were
used to modify states if necessary. This assumed that the following
request to the API actually succeeded, which is not always true. Now
that there is error handling for the requests, callback functions can
be called after successful book CRUD request. This way state updates are
done only if the request was successful.